+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Add `Info.immutable` to disable generating setters for public data members ([pull #461](https://github.com/bytedeco/javacpp/pull/461))
  * Map `String` to `char*` with `Charset.forName(STRING_BYTES_CHARSET)` when that macro is defined ([pull #460](https://github.com/bytedeco/javacpp/pull/460))
  * Fix `Loader.ClassProperties` not always getting overridden correctly when defined multiple times
  * Allow `Loader.load()` to also rename executables on extraction to output filenames specified with the `#` character